Adds completePurchase and updates the callback response
Skrill\Gateway::completePurchase now serves as an entrypoint for the
callbacks.

Skrill\Message\StatusCallback now properly validates the signatures and
takes the secretWord parameter passed into the base Gateway. It also
(attempts) to do some basic handling of this value; it'll accept both
the md5 hash and the secret word itself.